Get rid og gmail, use a normal imap/smtp server
Gmail was dangerous because of "security" which required frequent reauth
and a specific phone. This now use a normal imap/smtp server.

Also, no more need to save message ids in a file, just rely on imap SEEN
flag.

Save the state of affairs in a file (in simple json) so the program can
be interrupted and restarted with little damage.